Make certbundle.run container friendly
Running containerized services and applications in the context of
privileged users is a practice highly discouraged. This creates a
problem, given that said services and applications also require
a runtime container initialization stage during which container-wide
configuration changes are made, some of which include installing custom
certificates.

To solve this problem, some prefer using non-privileged user to install
custom certificates, which in the context of `ca-certificates` requires
changing the ownership of the `/var/lib/ca-certificates` folder and
running `update-ca-certificates` as a regular user. However, not all
of the `ca-certificates` scripts are designed to be executed by a
regular user.

This commit makes changes to the `certbundle.run` script to allow it
to be executed as a regular user (i.e. its previous form runs into
errors because newly created files have 444 permissions and a regular
user isn't allowed to modify them without explicitly setting the
owner write permission bit.